Move raider handler to break a dependency cycle
