+++ model/watched-member.js
@@ -78,6 +78,16 @@ const WatchedMember = {
         }
     },
 
+    raiderHandler: async (member) => {
+        const now = new Date();
+        const year = now.getFullYear();
+        const month = String(now.getMonth() + 1).padStart(2, '0');
+        const day = String(now.getDate()).padStart(2, '0');
+
+        await WatchedMember.add(member.id, `Suspected raid ${year}-${month}-${day}`);
+        Guild.watchlistChannel.send(trans('model.guild.raiderHandler', [member.toString()], 'en'));
+    },
+
     voiceStateUpdateHandler: (oldVoiceState, newVoiceState) => {
         if (WatchedMember.isMemberWatched(oldVoiceState.member.id)) {
             switch (true) {
